requirements.txt
openai
gradio
python-dotenv
fuzzywuzzy
plotly
markdown